TETO-103 Refactor: 프론트 포트번호 수정
