+++ backend/restaurant-service/build.gradle
@@ -1,3 +1,5 @@
+// ========= restaurant-service application.properties =========
+
 plugins {
     id 'java'
     id 'org.springframework.boot' version '3.5.6'
