+++ backend/restaurant-service/src/main/java/com/tabletopia/restaurantservice/domain/restaurant/entity/Restaurant.java
@@ -1,7 +1,7 @@
 package com.tabletopia.restaurantservice.domain.restaurant.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.tabletopia.restaurantservice.domain.restaurantcaCategory.entity.RestaurantCategory;
+import com.tabletopia.restaurantservice.domain.restaurantCategory.entity.RestaurantCategory;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.SQLDelete;
